+++ SampleApi/Library/Tax/Services/TaxService.cs
@@ -2,8 +2,8 @@ namespace Library.Tax.Calculator.Services
 {
   public interface TaxService
   {
-    EffectiveTaxRate? Calculate(float income, int year);
-    EffectiveTaxRate? Calculate(float income, int year, object response);
+    EffectiveTaxRate? Calculate(decimal income, int year);
+    EffectiveTaxRate? Calculate(decimal income, int year, object response);
     public object? QueryOnline(int year);
   }
